requirements.txt
pymupdf
python-docx
pytesseract
Pillow
streamlit
watchdog
